docker/py/app/requirements.txt
# pycryptodome
# websocket-client

fastapi[all]
uvicorn
pydantic

pandas
numpy
yfinance

psycopg2-binary
